v3.5 - New base code model, encrypted database, new name!
The entire framework of the module has been rebuilt from ground up to follow a proper state flow model.

Instead of plaintext, the database files are now serialized to JSON, compressed with GZIP, then prepended with the AES IV, encrypted using AES-256-CBC, and finally Base64-encoded. This makes them nearly impossible to decrypt on the drive and when they are loaded into the script, they are loaded into memory, but passwords still remain encrypted, until specifically retrieved and are then wiped from memory.

When a session is locked, variables are not just set to $null, they are overwritten. The key in memory is overwritten and wiped. Since it is a byte array, this is extremely effective, and the database is overwritten several times before all of the memory artefacts are set to $null. This makes it very difficult for memory forensics to capture any useful data.

After locking the session for inactivity, the script will also exit altogether, after another custom inactive session time.

The script now displays total entries and number of expired entries right on the main menu.

The add entry option has an advanced Paschword Generator that features several non-standard features, including the ability to customize it with a custom dictionary.

An extended help menu has been added.

F9 will provide details about the script's current configuration.